src/test/javascript/spec/app/entities/monitoreo/fotografia-monitoreo/fotografia-monitoreo.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { FotografiaMonitoreoService } from 'app/entities/monitoreo/fotografia-monitoreo/fotografia-monitoreo.service';
import { IFotografiaMonitoreo, FotografiaMonitoreo } from 'app/shared/model/monitoreo/fotografia-monitoreo.model';

describe('Service Tests', () => {
  describe('FotografiaMonitoreo Service', () => {
    let injector: TestBed;
    let service: FotografiaMonitoreoService;
    let httpMock: HttpTestingController;
    let elemDefault: IFotografiaMonitoreo;
    let expectedResult;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = {};
      injector = getTestBed();
      service = injector.get(FotografiaMonitoreoService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new FotografiaMonitoreo(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'image/png', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', async () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: elemDefault });
      });

      it('should create a FotografiaMonitoreo', async () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new FotografiaMonitoreo(null))
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should update a FotografiaMonitoreo', async () => {
        const returnedFromService = Object.assign(
          {
            nombre: 'BBBBBB',
            ruta: 'BBBBBB',
            extension: 'BBBBBB',
            fotografia: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should return a list of FotografiaMonitoreo', async () => {
        const returnedFromService = Object.assign(
          {
            nombre: 'BBBBBB',
            ruta: 'BBBBBB',
            extension: 'BBBBBB',
            fotografia: 'BBBBBB'
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a FotografiaMonitoreo', async () => {
        const rxPromise = service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
